Remove duplicate view and objc_allocateClassPair methods from Foundation Module
just a bit of cleanup, we had a view (fromObj) method in Foundation and a view module with a get method.   Now all the veiw specific method are in the View module.

Also, the view module had a duplicate of the Foundation external method bjc_allocateClassPair.  Its been deleted now.