Added shared methods to help dealing recognizers
recognizerByName finds a currently attached recognizer with a given name

removeRecognizers and removeRecognizersIncludingSubViews take a UIView ptr and remove all recognizers attached to that view (or that view and its subviews).  They are designed to help cleanup on pages where recognizers are attached implicitly.

More clean up of the Foundation and UIKit modules, most moving view related methods around and adding acouple Foundation methods (Foundation-isKindOf, and UIKit-IsDescendantOfView (to test for sub-views).